Update patch set 1
Patch Set 1:

The problem in 16457 isn't really that we're moving sqrt, it's that we're moving into a loop.  We need to teach tighten to not do that.

So I'm leaning towards not doing this.  We'd just have to undo it in 1.8 once we get a real fix.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: TryBot-Result=0