Update patch set 4
Patch Set 4:

Except for benchmarking to compare costs, export GO_SSA_PHI_LOC_CUTOFF=0 is the rule for all my development on this (I saw plenty of timing spew to let me know it was running).  And I am re-checking again to be certain.

This is good for submit into 1.7, to mitigate the heinous compilation time problem?

Patch-set: 4