Update patch set 1
Patch Set 1:

Thanks, Keith.

> Any function which could be the target of an indirect call would also need a stack check.

I'm not really sure how to detect that, particularly in the presence of unsafe, reflect, or assembly. I guess one could restrict the analysis to unexported functions in packages that don't import reflect or unsafe, when the -complete compiler flag is present, and which contain no indirect calls anywhere in the package. And you'd need to make sure that all call trees have roots in an exported function with a stack check. I'd bet that's a common set of features, but it sounds like a lot of machinery; I'm not sure it's worth it.

Maybe as a first pass for Go 1.8, I'll just extend this small fix to all architectures and add a comment about future possibilities. This CL seems to cut about 1% off the sort benchmarks, although it's noisy.

Patch-set: 1