Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> The assembly implementation looks good to me, although it may result in different numbers than the old backend does for numbers that do not fit in uint32. I guess it is fine by the spec?
> 
> An alternative way (safer but less efficient) would be doing float->int64->uint32 in the frontend.

Tried it, but it doesn't work.  The float->int64 conversions (in runtime/vlrt.go) depend on float->uint32 as a building block.  Infinite recursion ensues.

Maybe there's a way to fix the vlrt.go code to only use float->int32.  But that's a bucket of worms I'd rather not touch.

Patch-set: 2
Label: Code-Review=0