Update patch set 1
Patch Set 1:

I've changed the error type to a struct with a bit more context inside of it, cleared up the naming of the error variable/type, and removed the slice (pre)allocation.

I have added one extra test case exercising a mismatched host, but the current structure of the test does not allow me to assert anything about the error coming out of matchGoImport, just that it is not nil.

Is it a good idea to add more logic into the TestMatchGoImport function to assert that each test case returning an error matches a predefined error for that test?

Patch-set: 1
Reviewer: Gerrit User 14561 <14561@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0