[dev.garbage] runtime: start a ROC epoch
There are three basic operations involved in the ROC algorithm.
This CL provides startG. The other two are recycleG and publishG
and will be introduced in following CLs.

startG starts a new ROC epoch related to the currently running
goroutine. The epoch ends when the next epoch begins, when the
goroutine exits, or the cost of tracking the published objects
no longer warrants the expense.

Change-Id: I23a96aa994e3d84781c54ae1c7f0ee93f25afeb5