[dev.garbage] runtime: publish stack pointers for ROC
This is one of several CLs that call publish when
objects are being made available to another goroutine.

This publishes all the objects pointed to by the
goroutine's stack. While the routine is general it
currently is only used when publishing the stack of
a newly created goroutine. In particular in the
statement
	go someFun(ptrToLocal)
a new stack is created with a frame for the call to
someFun. The actual ptrToLocal is placed into the
frame on the new stack. This is potential a
publishing action since ptrToLocal may in fact
point to an unpublished local object in the
parent goroutine and it is being made available to
the child goroutine.

The implementation closely follows that of the scanstack
routine.

Change-Id: I845b97e62a4eef89de8101ae0fade258ae7f5160