[dev.garbage] runtime: publish finalizers for ROC
This is one of several CLs that call publish when
objects are being made available to another goroutine.

This publishes objects that are being finalized along
with the associated function / closure / free
variables. Publishing is required since when the finalizer
runs it will run on a different goroutine than the
one calling setFinalizer. Publishing the object being
finalized also ensures that the object will not be
recycled by ROC and all the logic for discovering and
running the finalizer will done as part of a normal GC
cycle.

Change-Id: I3752517660d151ea8576312c1cfe047c8c7a81e6