[dev.garbage] runtime: publish profile objects for ROC
This is one of several CLs that call publish when
objects are being made available to another goroutine.

This publishes the object that is being profiled since
it's offset off of the span's base is saved as part
of the spans special processing. Ultimately we need to teach
the profiler about ROC so it can report if ROC recycles the
object. By making the object public ROC will not recycle
the object even if it could and the profiler will report
information when the normal GC frees the object.

Change-Id: Ief0a9200430b8364962019ee6b5031065eb2b1f2