Update patch set 1
Patch Set 1:

R=go1.8

Note that the timing of the CL is a bit more complicated. For example, with jujud I see its build time drop from 24.5s to 23.3s with this CL. So I suspect it is acceptable as is because it helps with larger builds.

Also Ryan Brown was a Google Engineer when he wrote this CL, so it is owned by Google.

Patch-set: 1
Label: Code-Review=0