Update patch set 1
Patch Set 1:

(1 comment)

I have a new patch to upload with fixed comments, and a question.
For PPC, the loop-zero and loop-move pseudos work "best" with inputs that point one element before the stuff being moved, but those registers are clobbered.

I think there's a risk of an optimizer (stupidly) commoning/hoisting those even thought they are clobbered, and thus creating a potential GC problem, though this would be a "dumb" optimization to be avoided.  Do we declare that if the optimizer does this it is a bug to be fixed, or should we make the loop copy/zero code more defensive (at the cost of 1 or 2 instructions per zero or copy)?

Patch-set: 1