Update patch set 2
Patch Set 2: -Code-Review

We realized that "./bin/go test runtime" was NOT running the tests with ./bin/go, it was actually shelling out to the "go" on my $PATH which did not have this patch. With the correct "go" on my $PATH, the tests pass on amd64.

I filed #16577 to fix the runtime tests to use the right compiler.

Brad also asked me to test with GOARCH=386. With GOARCH=386 and GOHOSTARCH=386 we get either hangs or crashes in make.bash. Austin is looking at it.

Patch-set: 2
Label: Code-Review=0