Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1: Code-Review+2
> 
> (1 comment)
> 
> Should we consider this for more than just x86?

Maybe.  Do other archs need a temporary register to access globals in PIC mode?  I thought most had a reserved register for this purpose.

We're only doing this on x86 because it is hard to reserve CX across the board (like the old backend did) because 1) there aren't many registers, and 2) we have real instructions that require the use of CX.  I don't think that happens in any other architecture.

Patch-set: 1