encoding/json: add Decoder.DisallowUnknownFields option to reject unknown keys
It is not currently possible to perform strict input validation
in JSON unmarshaling. By design unrecognized (or colliding)
keys are silently ignored. This patch adds an option to
the Decoder that will collect all keys that do not have a
matching struct field. When this option is enabled the
Unmarshal() function emits an error when unused keys are detected.

The encoding/json unit tests have been refactored such that
each test case is run twice: once with DisallowUnknownFields() disabled
and once with DisallowUnknownFields() enabled. The majority of test
cases produce identical output under the two scenarios.
For the handful of test cases that produce different output,
the expected output or error has been updated in those test cases.

Fixes #15314

Change-Id: I3537848ffb002617a97a37b9d6f29833e0bcae8c