reflect, sort: add reflect.Swapper, sort.MakeInterface, sort.With
This CL adds two features that can be used separately or together.

The first feature is sort.MakeInterface, which returns a
sort.Interface given a length, swapper function, and less function.
Also included is a wrapper (sort.With) for brevity, equivalent to
calling sort.Sort(sort.MakeInterface(n, swapper, less)).

The second feature is reflect.Swapper which returns a specialized and
GC-compatible swapper function given a slice interface{}. The main
motivation of Swapper is so people can write:

    sort.With(len(s), reflect.Swapper(s), func(i, j int) bool {
        if s[i].Foo != s[j].Foo {
            return s[i].Foo < s[j].Foo
        }
        return s[i].Bar < s[j].Bar
    })

Because the calling conventions for interface calls differs from
function calls, sort.With cannot be implemented competitively if it's
used as an interface. This CL also recognizes the interface type
returned by sort.With and instead calls an auto-generated mirror of
the sort routines with the func calling convention.

As a result, even though it was not a goal, this CL actually makes
func-based sorting using sort.MakeInterface faster than
interface-based sorting:

	name                        time/op
	SortString1K-2              240µs ± 4%
	SortString1K_With-2         229µs ± 6%
	SortString1K_WithSwapper-2  224µs ± 5%

	StableInt1K-2               177µs ± 0%
	StableInt1K_With-2          134µs ± 0%
	StableInt1K_WithSwapper-2   140µs ± 0%

In the table above, the first of the three numbers is the traditional
sort. The second is a func-based sort (sort.With) with no
reflection. The last number is sort.With+reflect.Swapper. The second
and third numbers are effectively identical and in the noise. I only
ran with 10 iterations of each, though. I can do longer runs later.

DO NOT SUBMIT - Demo for #16721

Change-Id: I346530af1c5dee148ea9be85946fe08f23ae53e7