Update patch set 4
Patch Set 4: Code-Review+2

> Patch Set 4: -Code-Review
> 
> (1 comment)
> 
> So in BuildmodeShared, the linker runs twice, once to build an intermediate libstd.so (which uses -buildmode=shared), and once to build the binary that links against it (-linkshared).
> 
> If the model for BuildmodeShared is that libstd.so contains the DWARF info relevant for it, this is fine. If the model is that the final binary contains the DWARF info for both the binary and libstd.so, then eliding the go.info symbols here is not fine. Let me have a read of the linker and see if I can work out which it is...

Good (?) news, in BuildmodeShared DWARF output is disabled at the moment. In cmd/go/build.go, -w is passed in to the linker along with -linkshared. As there's no DWARF before this CL in that mode, there doesn't need to be any after this CL.

Patch-set: 4
Label: Code-Review=+2