Update patch set 9
Patch Set 9:

Ian: The long-term goal is to improve the style readability of the code. I also want to make it possible to unit-test parts of the linker by splitting the code into components. I think the best way to start is to remove many of the globals and make dependencies explicit.

Matthew: I made *ld.Link a receiver for the top level function calls in the main function and (mostly) kept it a parameter for the rest. I think most of the rest of the functions will eventually not take *ld.Link as a parameter but instead will be methods on smaller components.

I did this by hand, but tried to keep the changes as mechanical as possible so I could think through what a tool would need to do. I'm thinking of writing a tool that does something similar. I think it would also be useful to have a tool that can turn a parameter into a receiver and the reverse.

Patch-set: 9