cmd/compile: correct PPC64LE ld.so version; reduce stackguard
This fixes the version of ld.so specified by the linker
on PPC64LE, so that Go works again (make.bash produced a
non-running binary).  It is claimed that PPC64BE still uses
the old version, but lacking a machine to test I don't know
for sure.

The stackguard size is reduced slightly.  This is a result
of slightly optimized address generation for PPC, plus
(1) tweaking the runtime to move an error case out of
nosplit function and into a closure, and (2) replacing
a reference to a captured variable ("g") with a
recalculation of the same ("getg()") to get rid of a closure.

688 is better than 720, but still not as good as we want.

Updates: #16838, #16010, #9214

Change-Id: I7744fced9b27cf062c071d8f301b03e9a7ccb4f4