[dev.garbage] runtime: manage mcache structures for ROC
Each P has an associated mcache that is used to allocate
objects locally. Spans that are full but are still part
of a ROC epoch are linked using the span's nextUsedSpan
field. When a ROC epoch ends the spans in the mcache are
moved to either the list for spans with no more room in
them, spans with room for allocation that are not
completely empty, or onto a list of spans that have no
alloced objects in them and can be reused for any object
size.

Change-Id: I02b60dff42edfca0ddacb4c8a7cb4527313324cb