Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This CL seems to do a few different unrelated things, shouldn't it be split up?  Or maybe the PPC64LE work has not reached that stage?

It wasn't supposed to do two things, but then the output of ./make.bash failed to run, and that seemed like an important thing to fix.

PPC64 SSA is only just working, and this was supposed to claw back some of the stack guard bloat.  Claw back bloat is a combo of tweaking the library and adjusting StackGuard in 3 places, after Lynn Boger's CL that cut down on the number of temporaries generated for PPC.

Patch-set: 3