Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> (1 comment)

See "linux-arm builders/trybots seem to be hanging" on golang-dev, not sure how to link to it.  Ian thinks my change here is not the true fix, and instead merely hides the bug.

"The code to handle the cgo_dynamic_linker pragma is broken.  It no
longer correctly tests whether the flag was set.

It normally should not matter what the linker thinks the default
interpreter is, because for a cgo program it comes from
cgo_dynamic_linker and a non-cgo program is normally linked statically
and so the interpreter doesn't matter.  David's patch to set the right
interpreter is a good change but it covers up the actual bug.

Will send a fix later tonight."

Patch-set: 3