[dev.garbage] runtime: emit global to local write barriers
ROC needs to be informed when there is a pointer from
a global variable to a local object. Prior to ROC the
GC scanned all globals so this was not needed.
With ROC a local object needs to be made public
before it can be written into a global that can be seen
by another goroutine.

Change-Id: I4d568615c9104f94aade437eda42e4ecbba88c5f