Update patch set 5
Patch Set 5:

Aha! I know why there is a failing test. It is not an annihilation problem. There are two keys "x" and "X" that match to the same field when case-insensitive matching is performed. Now that I know why it's failing, and I've reproduced it in a simpler test case, I should be able to finish up this patch shortly.

Patch-set: 5
Label: Code-Review=0