Update patch set 1
Patch Set 1:

> Would it be possible to catch these situations with go vet? You would probably need a way to mark the structure as special, and then go vet could figure it out where a KeepAlive must be added.

I don't know.  Everything I've thought of so far could have quite a few false positives.  The cases that matter are where you have a parameter or local variable that is pointer to some type, there is or might be a finalizer on the object, the last reference to the object in the function is passing a field to some other function, and the finalizer might in some way invalidate that field value.

Patch-set: 1