Update patch set 3
Patch Set 3:

Had several thoughts this morning:

First, perhaps if it is expensive, we limit it to smaller functions (size TBD) where the garbage overhead would be smaller (or that could reuse memory from a cache), or only to those marked nosplit in the runtime because any wasted space there forces a stack guard increase which (currently) costs us per-thread on every platform.

Second, it could be quite platform-specific, because of variation in both code generation and platform-specific optimization.  PowerPC is currently the worst case for nosplit; I will play with it there.

Downside of the 2-algorithms strategy is of course increased testing obligations; that might argue against this, since stack bloat is annoying but not a showstopper like dozens-of-gigabyte compilations.

Patch-set: 3