Update patch set 1
Patch Set 1:

> 1) I would put the whitelist files one dir down

Good idea. Will do.

> 2) Maybe allow a * at the end of the line (that's all) to allow matches to be shorter. Or, perhaps better, make the text need only be the prefix of the error, not the whole thing.

We omitted line numbers from these messages to reduce churn. But in exchange, the assumption was that the error messages would be distinct enough to uniquely identify the failures. I'd rather leave the whole line.

Using the full line also simplifies the implementation and mental model a fair amount. (What if one prefix matches multiple lines? Do we pick the first one? What if a line with a longer prefix needed that one to match, and we should have picked a different one?)

> 3) There should be a tool to recreate these lists if required. That's going to be hard because you've added comments, but as I see it now this state is unwieldy and unmaintainable. If someone improves an error message...

That's where I started. But the platform-specific variation in the errors drove me this direction. This is perhaps less easy to maintain, but much easier to grok as a human, even before comments.

There are several hundred errors and that's after deduplication and templating. I think that writing a tool to automatically categorize these errors to generate/maintain a similar minimal set of files is overkill. But writing out a complete 400 line list of errors for every single platform also seems over the top.

I hear you about the pain of improving an error message, but I suspect that less than a pound of of sed is enough cure, and that the ounce of prevention here is pretty heavy.

I'm open to input about *how* to autogenerate these, and I'm willing to lose the comments. (They can always be moved to a README anyway.)

Patch-set: 1