Update patch set 2
Patch Set 2:

> What's the status of s390x SSA? If that's done, then you could
 > maybe intrinsify these and get rid of the call overhead too.
 > 
 > See https://github.com/golang/go/issues/13095#issuecomment-242785555

It's not quite fully functional yet, but close. I hope to start sending CLs fairly soon.

I'm definitely going to try and implement the current intrinsics. I'd be happy to do some of the math functions too if there is broad agreement it's desirable. I suspect it would make sense to inline math.Sqrt in the SSA backend rather than treating it as a special case elsewhere in the compiler.

Patch-set: 2
Label: Code-Review=0