cmd/internal/obj, cmd/link: add relocation variant to object file
DO NOT SUBMIT - proposal.

This commit modifies the object file format. Tools that read object
files (e.g. stringer) will need to be recompiled after this change.

Relocation variants can be useful for specifying relocations that
differ only in the way they encode values. For example, ppc64 uses
relocation variants to specify if the high or low bits of a value
should be encoded.

This commit adds relocation variants to the object file format so
that they can be passed from the compiler to the linker. Before
this change a unique relocation was required for each type/variant
combination resulting in complexity in the linker.

For now I've only modified the s390x linker to take advantage of
this change, however ppc64 and perhaps other architectures
could benefit from using it too.

On s390x this commit increases the size of the .a files in
$GOROOT/pkg by an average of 1.6%, with the maximum increase being
3.2%. It will most likely increase the size of object files on
other platforms by a similar amount even if they don't use
relocation variants.

Fixes #14218.

Change-Id: I2f2371e64d9e7fc884c67af40f0e6dc6dc28f731