Update patch set 1
Patch Set 1:

So what happened here is that in Go 1.4 and before, the linker generated a bunch of symbols directly into the runtime package's namespace. In Go 1.5 I changed this as part of -buildmode=shared support to move all this data into a structure (moduledata) so that we could have a linked list of them, one per shared object. But, out of possibly misplaced paranoia, I left the runtime.xxx symbols in case there were external tools that used these symbols to analyse executables (e.g. src/cmd/internal/objfile/pe.go references runtime.pclntab).

This should all be in a comment in the source somewhere, I guess.

Patch-set: 1