Update patch set 1
Patch Set 1:

I think at a minimum the package global symt symbol should be moved inside the method where it's being used.

There's also another problem which is the way that by asking for the symbol, we actaully generate it, then using a reference to the symbol stored in a private cache, we monkey with some of it's configuration, and later on, pe.go and elf.go do some more monkeying.

This is kind of spooky action at a distance, ordering matters, as well as who creates the symbol, by side effect, and when.

It almost feels like there should be a NewSym(sym *Symbol) method that _inserts_ the symbol into the map, explicitly, deliberately, and immutably.

Patch-set: 1