Update patch set 1
Patch Set 1:

I'm not too excited about this.  We've already got an int32 field for the reloc type, and I count 42 different reloc types.  We're not exactly in danger of running out of bits.  What's the big advantage to adding another int32 field?

If the variant is really useful, what if we just reserve 8 bits in the current int32 field to hold the variant?

Patch-set: 1