Update patch set 1
Patch Set 1:

> I'm not too excited about this.  We've already got an int32 field for the reloc type, and I count 42 different reloc types.  We're not exactly in danger of running out of bits.  What's the big advantage to adding another int32 field?
> 
> If the variant is really useful, what if we just reserve 8 bits in the current int32 field to hold the variant?

Yeah, we can reuse some of the bits from the current int32 field.

The goal of reloc variant is to group similar relocation types together
so that most (semi-)target independent code doesn't need to care about
detailed instruction encoding for a specific relocation (only the
target specific linker needs to concern with those details).

For example, Power uses two different relocations for addresses because
the constants are encoded differently. I think right now the linker
actually read the raw instruction bytes to determine which encoding
to use. If we have variants, then we can have a single R_ADDRPOWER
and two variants (D-form and DS-form), so that we keep the number of
relocation type constant but we remove the ugly hack in the power
linker.

I'm fine with reducing the size of RelocType to int16 and introduce
an int16 for the relocation variant.

Patch-set: 1
Label: Code-Review=0