Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > This is okay, but also a little sad that the compiler doesn't just
>  > deal with this.
>  > 
>  > I found the old code marginally more clear.
>  > 
>  > Can you file or find a bug about this and reference this CL?
>  > 
>  > (But submitting this is fine.)
> 
> Sure thing. 
> 
> Personally I'm not sure it's the compilers job to make
> 
>      p := new(T)
>      var pp *T
>      *pp = *p
> 
> and
> 
>      p := new(T)
>      pp := *p
> 
> equal.

I think it is, because visibility-wise to the user there's no difference, at least in the case where p is provably non-nil. The clear is just redundant.

I actually think we'll get that for free from the existing compiler as some of the runtime call lowering is removed from the frontend and moved into the responsibility of the backend.

But I agree we shouldn't go out of our way for it. I mostly want to track it because I do think it'll just fall out naturally.

Patch-set: 3