cmd/compile: handle pragmas immediately with -newparser=1
Instead of saving all pragmas and then processing them after parsing
is finished, process them immediately during scanning like the current
lexer does.

This is a bit unfortunate because it means we can't use
syntax.ParseFile to concurrently parse files yet, but it fixes how we
report syntax errors in the presence of //line pragmas.

Change-Id: I3a3fcd2ef2efac921ab6cc5c0866c7099def0747