compress: add generator format for compression tests
We add support for a very lightweight format (called bitgen)
for constructing bitstreams from a more descriptive syntax.
This allows us to manually craft valid and malicious streams
to thoroughly test decompressors. This allows the testdata to
be more self-describing than just a hex-string that is
uninformative about itself.

The intent is to use this format to thoroughly test flate.Reader
and bzip2.Reader.

Using the bitgen format, we create a new test TestReader that consolidates
the test functionality of the following:
	TestStreams
	TestIssue5915
	TestIssue5962
	TestTruncatedStreams
	TestReaderTruncated
	TestNlitOutOfRange

Change-Id: I7950faca831802eee678b68522f74a83db9fc7a0