Update patch set 2
Patch Set 2:

> Shouldn't those all be inlined and CSE'd?

Good question.

They are all inlined. However, it is possible to reach !hbits.isPointer(), line 1157, without evaluating !hbits.morePointers(), line 1154, if i != 1*sys.PtrSize is false. So they're not safe to CSE. (In SSA speak, neither of the values is in a block that in an ancestor of the other.) You can get them to CSE by flipping the order of the expressions in line 1154, but that still generates inferior code to this CL...and instead of the intent being obvious and explicit, it's subtle and prone to breakage.

This CL does mean that we needlessly load hbits in some cases (otherwise, indeed, the compiler would have done this for us), but that's (apparently) an acceptable trade-off.

Patch-set: 2