net: fix parsing of negative decimals
Commit 00b779aeed9e0345e4bdbb38c85df63a6352b954 fixed the dtoi function such
that it now properly parses negative number. Ironically, this causes several other
functions that depended on dtoi to now (incorrectly) parse negative numbers.

For example, ParseCIDR("-1.0.0.0/32") used to be rejected prior to the above
commit, but is now accepted even though it is an invalid CIDR notation.
This CL fixes that regression.

We seperate the negative parsing logic out as utoi (unsigned decimal to integer)
and use that in most places where dtoi was otherwise used. The changes made
assumes that port numbers, protocol numbers, and zone indexes are always unsigned.

The also fixes a minor bug where ParseCIDR("0.0.0.0/-0") used to pass, but
really should be rejected.

Updates #16350

Change-Id: I414bb1aa27d0a226ebd4b05a09cb40d784691b43